Fix integrity check since #8070 (#8137)
* Fix integrity check since #8070

* Add Comment

+++ protected/humhub/modules/like/Events.php
@@ -53,17 +53,23 @@ class Events extends BaseObject
         $integrityController = $event->sender;
         $integrityController->showTestHeadline("Like (" . Like::find()->count() . " entries)");
 
+        /** @var Like $like */
         foreach (Like::find()->each() as $like) {
-            if ($like->source === null) {
-                if ($integrityController->showFix("Deleting like id " . $like->id . " without existing target!")) {
-                    $like->delete();
-                }
+
+            // Check underlying record exists
+            if (
+                !$like->getContent()->exists()
+                && $integrityController->showFix("Deleting like id " . $like->id . " without existing target!")
+            ) {
+                $like->delete();
             }
+
             // User exists
-            if ($like->user === null) {
-                if ($integrityController->showFix("Deleting like id " . $like->id . " without existing user!")) {
-                    $like->delete();
-                }
+            if (
+                !$like->getContent()->exists()
+                && $integrityController->showFix("Deleting like id " . $like->id . " without existing user!")
+            ) {
+                $like->delete();
             }
         }
     }
